requirements.txt
aiogram==3.1.1
asyncpg==0.29.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
aiohttp==3.9.1